feat(api): don't set reactnative's driver as default & allow missing driver
This lets us use the plugin in any environement (react-native / web / node.js) since AsyncStorage is not required from the plugin's code anymore.

This is a breaking change from versions 2.x since you now need to set  the driver explicitly in the API configuration.

You can also use the wrapper without the caching features if you want to. This can be useful in isomorphic applications where you don't want the server code to have any sort of caching (or if you simply don't have a driver for it).